requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
crispy-bootstrap4==2022.1
cryptography==39.0.1
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.7
django-allauth==0.52.0
django-crispy-forms==2.0
django-jazzmin==2.6.0
django-smtp-ssl==1.0
gunicorn==20.1.0
idna==3.4
lighthouseweb3==0.1.4
oauthlib==3.2.2
Pillow==9.4.0
protobuf==3.20.3
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.8
python-dotenv==0.21.0
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
urllib3==1.26.14
whitenoise==6.2.0
